requirements.txt
asgiref==3.8.1
black==24.4.2
bleach==6.1.0
click==8.1.7
DateTime==5.5
Django==5.0.7
django-cors-headers==4.4.0
django-summernote==0.8.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
webencodings==0.5.1
zope.interface==6.4.post2